test: fix test case
