+++ packages/react-server/examples/basic/src/routes/test/transition/_client.tsx
@@ -2,7 +2,6 @@
 
 import { Link, useRouter } from "@hiogawa/react-server/client";
 import { cls } from "../../../components/utils";
-import { changeCounter } from "./_action";
 
 const TABS = [
   ["/test/transition", "About"],
@@ -30,33 +29,5 @@ export function Tablist() {
         </Link>
       ))}
     </ul>
-  );
-}
-
-export function Counter(props: { value: number }) {
-  const isActionPending = useRouter((s) => s.isActionPending);
-
-  return (
-    <form action={changeCounter} className="flex flex-col items-start gap-2">
-      <div className="flex items-center gap-2">
-        <button
-          className="antd-btn antd-btn-default px-2"
-          name="delta"
-          value={-1}
-        >
-          -1 (2.0 sec)
-        </button>
-        <button
-          className="antd-btn antd-btn-default px-2"
-          name="delta"
-          value={+1}
-        >
-          +1 (0.2 sec)
-        </button>
-        <div className={cls(isActionPending && "opacity-50")}>
-          Count: {props.value}
-        </div>
-      </div>
-    </form>
   );
 }
