+++ packages/react-server/examples/basic/src/routes/test/transition/layout.tsx
@@ -1,6 +1,5 @@
 import type { LayoutProps } from "@hiogawa/react-server/server";
-import { getCounter } from "./_action";
-import { Counter, Tablist } from "./_client";
+import { Tablist } from "./_client";
 
 // similar demo
 // https://react.dev/reference/react/useTransition#marking-a-state-update-as-a-non-blocking-transition
@@ -12,10 +11,6 @@ export default async function Layout(props: LayoutProps) {
         <h4 className="font-bold">Navigation State</h4>
         <Tablist />
         <div className="p-2">{props.children}</div>
-      </div>
-      <div className="border p-3 flex flex-col gap-2">
-        <h4 className="font-bold">Action state</h4>
-        <Counter value={getCounter()} />
       </div>
     </div>
   );
